+++ docs/.vuepress/config.js
@@ -53,7 +53,7 @@ module.exports = ctx => ({
       }
     ]
   ],
-  theme: '@vuepress/vue',
+  // theme: '@vuepress/vue',
   plugins: [
     [
       '@vuepress/pwa',
@@ -70,6 +70,11 @@ module.exports = ctx => ({
           indexName: 'vue-router'
         }
       : null,
+    carbonAds: {
+      carbon: 'CEBICK3I',
+      custom: 'CEBICK3M',
+      placement: 'routervuejsorg'
+    },
     repo: 'vuejs/vue-router',
     docsDir: 'docs',
     smoothScroll: true,