+++ examples/hash-mode/app.js
@@ -35,6 +35,7 @@ const Home = { template: '<div>home</div>' }
 const Foo = { template: '<div>foo</div>' }
 const Bar = { template: '<div>bar</div>' }
 const Unicode = { template: '<div>unicode: {{ $route.params.unicode }}</div>' }
+const Query = { template: '<div>query: "{{ $route.params.q }}"</div>' }
 
 // 3. Create the router
 const router = new VueRouter({
@@ -45,7 +46,8 @@ const router = new VueRouter({
     { path: '/foo', component: Foo },
     { path: '/bar', component: Bar },
     { path: '/é', component: Unicode },
-    { path: '/é/:unicode', component: Unicode }
+    { path: '/é/:unicode', component: Unicode },
+    { path: '/query/:q', component: Query, name: 'param' }
   ]
 })
 
@@ -66,14 +68,17 @@ const vueInstance = new Vue({
         <li><router-link to="/é/ñ">/é/ñ</router-link></li>
         <li><router-link to="/é/ñ?t=%25ñ">/é/ñ?t=%ñ</router-link></li>
         <li><router-link to="/é/ñ#é">/é/ñ#é</router-link></li>
+        <li><router-link to="/query/A%25">/query/A%</router-link></li>
+        <li><router-link :to="{ name: 'param', params: { q: 'A%' }}">/query/A% (object)</router-link></li>
+        <li><router-link to="/query/A%2FE">/query/A%2FE</router-link></li>
+        <li><router-link :to="{ name: 'param', params: { q: 'A/E' }}">/query/A%2FE (object)</router-link></li>
       </ul>
       <pre id="query-t">{{ $route.query.t }}</pre>
       <pre id="hash">{{ $route.hash }}</pre>
       <router-view class="view"></router-view>
     </div>
   `,
-  methods: {
-  }
+  methods: {}
 }).$mount('#app')
 
 document.getElementById('unmount').addEventListener('click', () => {