+++ src/create-matcher.js
@@ -175,6 +175,14 @@ function matchRoute (
   path: string,
   params: Object
 ): boolean {
+  try {
+    path = decodeURI(path)
+  } catch (err) {
+    if (process.env.NODE_ENV !== 'production') {
+      warn(false, `Error decoding "${path}". Leaving it intact.`)
+    }
+  }
+
   const m = path.match(regex)
 
   if (!m) {
@@ -185,10 +193,9 @@ function matchRoute (
 
   for (let i = 1, len = m.length; i < len; ++i) {
     const key = regex.keys[i - 1]
-    const val = typeof m[i] === 'string' ? decodeURIComponent(m[i]) : m[i]
     if (key) {
       // Fix #1994: using * with props: true generates a param named 0
-      params[key.name || 'pathMatch'] = val
+      params[key.name || 'pathMatch'] = m[i]
     }
   }