+++ src/create-route-map.js
@@ -85,6 +85,7 @@ function addRouteRecord (
     regex: compileRouteRegex(normalizedPath, pathToRegexpOptions),
     components: route.components || { default: route.component },
     instances: {},
+    enteredCbs: {},
     name,
     parent,
     matchAs,